Troy: Fixing this to use the new FBJS API
Summary:
Troy should now compile in a fresh checkout which is refreshing. And as a bonus it produces human-readable output!

Reviewed By: dweatherford

Test Plan: I ran externalize on 'photo-tag' and that output a really big JS file that didn't have any syntax errors in it, so it's probably good.

Revert Plan: ok