Fix ambiguities between ObjectLiteral & Block
Summary:
There's been a long standing bug in the parser in that it can't properly parse `if(1){}`. The problem was an ambiguity between ObjectLiteral & Block. The ES-262 spec addresses this by saying an ExpressionStatement may not begin with a t_LCURLY or t_FUNCTION, though it's been unclear how to build this into the grammar. The answer was to duplicate the Expression grammar again, except removing all possible reductions that begin with t_LCURLY or t_FUNCTION (i.e. ObjectLiteral or FunctionExpression). The ES-262 spec does something similar to handle ambiguities with the in operator in for(in) loops.

The new grammar has no reduce\reduce or shift\reduce conflicts so I've removed %glr-parser.

This also removes an unintentional `using namespace fbjs` in parser.yacc.hpp (it was generated, so it wasn't quite obvious). Also the `using namespace std` in node.hpp.

Reviewed By: dweatherford

Test Plan:

Revert Plan: ok