Add special-casing for `arguments`
Summary:
If the developer uses the `arguments` object in a function we need to catch this and run it through our runtime first. We only do this in functions that reference `arguments` to avoid the runtime cost if they're not going to need it.

Reviewed By: dweatherford

Test Plan: compiles

Revert Plan: ok