Apply licences, fix build bugs, and some stuff to the readme.
Summary:
Dave missed a few spots for the licenses, so I'm getting those here. I scp'd this to a machine which has an up-to-date copy of gcc and fixed some build errors we didn't see on FC4. Also added a little bit of information to the readme.

Reviewed By: epriestley

Test Plan: make fbjs jsbeautify

Revert Plan: ok