js intl: verify tx() calls in c++ binary
Summary:
The php js parser is slow and prohibits, for example, running the parser as part of analyze_resources. Porting the parser to C++ makes the parsing two orders of magnitude faster.

For this first round, just using the parser to verify that developers are using tx() correctly. The integration with analyze_resources is coming in a separate diff. Can also include it as part of an svn hook.

Eventually the php parser will be completely deprecated in favor of this C++ binary (although the parser has problems with things like function() { a({})} which needs to be fixed before going forward).

Reviewed By: marcel

Test Plan:
1. ran parser against all js files html/js and all checked out ok.
     2. inserted a bad call like tx(lklots) and correct, helpful error message cameup

Revert: OK

DiffCamp Revision: 54296