bypass libfbjs segfault on empty if(){}
Summary:
Got a segfault running jsxmin based of the parser in fbcode/libfbjs/ with head at 12964. A dynamic_cast was failing because the children nodes of an if statement are being populated incorrectly when the body of an if statement is empty.
Codez

  if (whatever) {

  }

Stack trace

  > #1  0x0000000000411d1c in fbjs::NodeIf::render (this=0x6b3b70,
  > guts=0x7fff31d312f0, indentation=1) at node.cpp:978
  > 978         while (firstStatement &&
  > dynamic_cast<NodeStatementList*>(firstStatement) != NULL

I doubt that this is the 100% correct fix but it works.

Reviewed By: marcel

Test Plan:
build ran jsxmin on test case and worked ok. ran jsxmin against all files no more segfaults.

Revert: OK

DiffCamp Revision: 65907