NodeWalker and removal of reduce()
Summary:
A general solution is needed for walking these AST's and applying a transformation to them. This implements a typical visitor class from which you can extend to subscribe to nodes you want to handle. There's built-in methods for manipulating the tree as you go, and getting data from your children.

Test Plan:
Built a few walkers for another application, adapting this API as I found use cases.

Reviewed By: epriestley